LLM_Assistant_MiniC/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
pyserial==3.5
